+++ keyboards/niu_mini/keymaps/readme.md
@@ -21,3 +21,4 @@ When adding your keymap to this list, keep it organised alphabetically (select l
 - **default** default layout from KBDFans
 - **mason**
 - **planck** Planck default layout
+- **xtonhasvim** A Planck-like layout with a few tweaks and a vim emulation layer.
