+++ users/xtonhasvim/rules.mk
@@ -1,0 +1,1 @@
+SRC += xtonhasvim.c
