+++ users/xtonhasvim/xtonhasvim.c
@@ -1,27 +1,4 @@
- /* Copyright 2015-2017 Jack Humbert
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include QMK_KEYBOARD_H
-#include "action_layer.h"
-
-LEADER_EXTERNS();
-
-#define X_____X KC_NO
-
-extern keymap_config_t keymap_config;
+#include "xtonhasvim.h"
 
 /************************************
  * helper foo
@@ -30,199 +7,35 @@ extern keymap_config_t keymap_config;
 #define PRESS(kc) register_code(kc)
 #define RELEASE(kc) unregister_code(kc)
 
-void TAP(uint16_t keycode) {
+static void TAP(uint16_t keycode) {
     PRESS(keycode);
     RELEASE(keycode);
 }
 
-void CMD(uint16_t keycode) {
+static void CMD(uint16_t keycode) {
   PRESS(KC_LGUI);
     TAP(keycode);
   RELEASE(KC_LGUI);
 }
 
-void CTRL(uint16_t keycode) {
+static void CTRL(uint16_t keycode) {
   PRESS(KC_LCTRL);
     TAP(keycode);
   RELEASE(KC_LCTRL);
 }
 
-void SHIFT(uint16_t keycode) {
+static void SHIFT(uint16_t keycode) {
   PRESS(KC_LSHIFT);
     TAP(keycode);
   RELEASE(KC_LSHIFT);
 }
 
-void ALT(uint16_t keycode) {
+static void ALT(uint16_t keycode) {
   PRESS(KC_LALT);
     TAP(keycode);
   RELEASE(KC_LALT);
 }
 
-/************************************
- * states
- ************************************/
-
-enum layers {
-  _QWERTY,
-  _LOWER,
-  _RAISE,
-  _ADJUST,
-  _MOVE,
-  _MOUSE,
-  // vim layers
-  _EDIT,
-  _CMD
-};
-
-enum keycodes {
-  DUMMY = SAFE_RANGE,
-  VIM_START, // bookend for vim states
-  VIM_A,
-  VIM_B,
-  VIM_C,
-  VIM_CI,
-  VIM_D,
-  VIM_DI,
-  VIM_E,
-  VIM_H,
-  VIM_G,
-  VIM_I,
-  VIM_J,
-  VIM_K,
-  VIM_L,
-  VIM_O,
-  VIM_P,
-  VIM_S,
-  VIM_U,
-  VIM_V,
-  VIM_VS, // visual-line
-  VIM_VI,
-  VIM_W,
-  VIM_X,
-  VIM_Y,
-  VIM_PERIOD, // to support indent/outdent
-  VIM_COMMA,  // and toggle comments
-  VIM_SHIFT, // avoid side-effect of supporting real shift.
-  VIM_ESC // bookend
-};
-
-/************************************
- * keymaps!
- ************************************/
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-/* Qwerty
- * ,-----------------------------------------------------------------------------------.
- * | Tab  |   Q  |   W  |   E  |   R  |   T  |   Y  |   U  |   I  |   O  |   P  | Bksp |
- * |------+------+------+------+------+-------------+------+------+------+------+------|
- * | Ctrl |   A  |   S  |   D  |   F  |   G  |   H  |   J  |   K  |   L  |  [;] |  "   |
- * |------+------+------+------+------+------|------+------+------+------+------+------|
- * | Shift|   Z  |   X  |   C  |   V  |   B  |   N  |   M  |   ,  |   .  |   /  |Enter |
- * |------+------+------+------+------+------+------+------+------+------+------+------|
- * | Brite|  Esc | Alt  | GUI  |Lower |    Space    |Raise | Left | Down |  Up  |Right |
- * `-----------------------------------------------------------------------------------'
- */
-[_QWERTY] = {
-  {KC_TAB,  KC_Q,           KC_W,    KC_E,    KC_R,          KC_T,    KC_Y,    KC_U,          KC_I,    KC_O,    KC_P,              KC_BSPC},
-  {LCTL_T(KC_ESC), LT(_MOVE,KC_A), KC_S,    KC_D,    KC_F,          KC_G,    KC_H,    KC_J,          KC_K,    KC_L,    LT(_MOVE,KC_SCLN), KC_QUOT},
-  {KC_LSFT, KC_Z,           KC_X,    KC_C,    KC_V,          KC_B,    KC_N,    KC_M,          KC_COMM, KC_DOT,  KC_SLSH,   RSFT_T(KC_ENT) },
-  {TG(_MOUSE), VIM_START,  KC_LALT, KC_LGUI, OSL(_LOWER),   KC_SPC,  KC_SPC,  OSL(_RAISE),   KC_LGUI, KC_LALT, X_____X,   TG(_MOVE)}
-},
-
-/* Lower
- * ,-----------------------------------------------------------------------------------.
- * |  ~   |  F1  |  F2  |  F3  |  F4  |  F5  |  F6  |   _  |   +  |   {  |   }  | Bksp |
- * |------+------+------+------+------+-------------+------+------+------+------+------|
- * |  Del |   !  |   @  |   #  |   $  |   %  |   ^  |   &  |   *  |   (  |   )  |  |   |
- * |------+------+------+------+------+------|------+------+------+------+------+------|
- * |      |  F7  |  F8  |  F9  |  F10 |  F11 |  F12 |ISO ~ |ISO | | Home | End  |      |
- * |------+------+------+------+------+------+------+------+------+------+------+------|
- * |      |      |      |      |      |             |      | Next | Vol- | Vol+ | Play |
- * `-----------------------------------------------------------------------------------'
- */
-[_LOWER] = {
-  {KC_TILD,  KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_F6,   KC_UNDS,    KC_PLUS,    KC_LCBR, KC_RCBR, KC_BSPC},
-  {KC_DEL, KC_EXLM, KC_AT,   KC_HASH, KC_DLR,  KC_PERC, KC_CIRC, KC_AMPR,    KC_ASTR,    KC_LPRN, KC_RPRN, KC_PIPE},
-  {_______, KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,  _______, _______, KC_HOME, KC_END,  _______},
-  {_______, _______, _______, _______, _______, _______, _______, OSL(_ADJUST),    KC_MNXT,    KC_VOLD, KC_VOLU, KC_MPLY}
-},
-
-/* Raise
- * ,-----------------------------------------------------------------------------------.
- * |   `  |  F1  |  F2  |  F3  |  F4  |  F5  |  F6  |   -  |   =  |   [  |   ]  | Bksp |
- * |------+------+------+------+------+-------------+------+------+------+------+------|
- * |  Del |   1  |   2  |   3  |   4  |   5  |   6  |   7  |   8  |   9  |   0  |  \   |
- * |------+------+------+------+------+------|------+------+------+------+------+------|
- * |      |  F7  |  F8  |  F9  |  F10 |  F11 |  F12 |ISO # |ISO / |Pg Up |Pg Dn |      |
- * |------+------+------+------+------+------+------+------+------+------+------+------|
- * |      |      |      |      |      |             |      | Next | Vol- | Vol+ | Play |
- * `-----------------------------------------------------------------------------------'
- */
-[_RAISE] = {
-  {KC_GRV,  KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_F6,   KC_MINS, KC_EQL,  KC_LBRC, KC_RBRC, KC_BSPC},
-  {KC_DEL,  KC_1,    KC_2,    KC_3,    KC_4,    KC_5,    KC_6,    KC_7,    KC_8,    KC_9,    KC_0,    KC_BSLS},
-  {_______, KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,  _______, _______, KC_PGUP, KC_PGDN, _______},
-  {_______, _______, _______, _______, OSL(_ADJUST), _______, _______, _______, KC_MNXT, KC_VOLD, KC_VOLU, KC_MPLY}
-},
-
-
-/* Adjust (Lower + Raise)
- * ,-----------------------------------------------------------------------------------.
- * |      | Reset|      |      |      |      |      |      |      |      |      |  Del |
- * |------+------+------+------+------+-------------+------+------+------+------+------|
- * |      |      |      |Aud on|Audoff|AGnorm|AGswap|Qwerty|Colemk|Dvorak|Plover|      |
- * |------+------+------+------+------+------|------+------+------+------+------+------|
- * |      |Voice-|Voice+|Mus on|Musoff|MIDIon|MIDIof|      |      |      |      |      |
- * |------+------+------+------+------+------+------+------+------+------+------+------|
- * |      |      |      |      |      |             |      |      |      |      |      |
- * `-----------------------------------------------------------------------------------'
- */
-[_ADJUST] = {
-  {RGB_MODE_PLAIN, RESET,   DEBUG,   _______, _______, _______, _______, _______, _______, _______, _______, KC_DEL },
-  {RGB_MODE_REVERSE, _______, MU_MOD,  AU_ON,   AU_OFF,  AG_NORM, AG_SWAP, _______, _______, _______, _______, RGB_VAI},
-  {RGB_MODE_FORWARD, MUV_DE,  MUV_IN,  MU_ON,   MU_OFF,  MI_ON,   MI_OFF,  _______, _______, _______, _______, RGB_VAD},
-  {RGB_TOG, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______}
-},
-
-
-/* movement layer (hold semicolon)
- */
-[_MOVE] = {
-  {TO(_QWERTY), X_____X, X_____X, X_____X, X_____X, X_____X, KC_HOME, KC_PGDN, KC_PGUP, KC_END, X_____X, X_____X},
-  {_______,     X_____X, LGUI(KC_LBRC), LGUI(LSFT(KC_LBRC)), LGUI(LSFT(KC_RBRC)), LGUI(KC_RBRC), KC_LEFT, KC_DOWN, KC_UP,   KC_RGHT, X_____X, X_____X},
-  {_______,     X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, _______},
-  {X_____X,     _______, _______, _______, _______, X_____X, X_____X, _______, _______, _______, TO(_QWERTY), _______}
-},
-
-/* mouse layer
- */
-[_MOUSE] = {
-  {TO(_QWERTY), X_____X, X_____X, KC_MS_UP, X_____X, X_____X, KC_MS_WH_LEFT, KC_MS_WH_DOWN, KC_MS_WH_UP, KC_MS_WH_RIGHT, X_____X, X_____X  },
-  {_______,     X_____X, KC_MS_LEFT, KC_MS_DOWN, KC_MS_RIGHT, X_____X, X_____X, KC_MS_BTN1, KC_MS_BTN2, KC_MS_BTN3, X_____X, X_____X},
-  {_______,     X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, X_____X, _______},
-  {_______,     _______, _______, _______, _______, X_____X, X_____X, _______, _______, _______, TO(_QWERTY), _______}
-},
-
-/* vim edit mode. just has an escape -> _CMD key */
-[_EDIT] = {
-  {_______,     _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______},
-  {VIM_START,   _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______},
-  {_______,     _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______},
-  {_______,     _______, _______, _______, _______, _______, _______, _______, _______, _______, TO(_QWERTY), _______}
-},
-
-/* vim command layer.
- */
-[_CMD] = {
-  {X_____X, X_____X, VIM_W, VIM_E, X_____X, X_____X, VIM_Y, VIM_U, VIM_I, VIM_O, VIM_P, X_____X},
-  {VIM_ESC,    VIM_A, VIM_S, VIM_D, X_____X, VIM_G, VIM_H, VIM_J, VIM_K, VIM_L, X_____X, X_____X},
-  {VIM_SHIFT,     X_____X, VIM_X, VIM_C, VIM_V, VIM_B, X_____X, X_____X, VIM_COMMA, VIM_PERIOD, X_____X, VIM_SHIFT},
-  {X_____X,     X_____X, _______, _______, X_____X, X_____X, X_____X, X_____X, _______, _______, TO(_QWERTY), X_____X}
-}
-
-};
 
 uint16_t vstate = VIM_START;
 bool yank_was_lines = false;
@@ -280,7 +93,7 @@ void simple_movement(uint16_t keycode) {
   }
 }
 
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+bool process_record_xtonhasvim(uint16_t keycode, keyrecord_t *record) {
   if(record->event.pressed && layer_state_is(_CMD) && IS_MOD(keycode)) {
     mod_override_layer_state = layer_state;
     mod_override_triggering_key = keycode;
@@ -762,60 +575,4 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
   } else {
     return true;
   }
-}
-
-/*
-  _QWERTY,
-  _LOWER,
-  _RAISE,
-  _ADJUST,
-  _MOVE,
-  _MOUSE,
-  // vim layers
-  _EDIT,
-  _CMD
-  */
-
-void rgbflag(uint8_t r, uint8_t g, uint8_t b) {
-  for(int i = 0; i < RGBLED_NUM; i++){
-    switch(i) {
-      case 9 ... 12:
-        // rgblight_setrgb_at(r,g,b,i);
-        led[i].r = r;
-        led[i].g = g;
-        led[i].b = b;
-        break;
-      default:
-        // rgblight_setrgb_at(0,0,0,i);
-        led[i].r = 0;
-        led[i].g = 0;
-        led[i].b = 0;
-        break;
-    }
-  }
-  rgblight_set();
-}
-
-uint32_t layer_state_set_user(uint32_t state) {
-    switch (biton32(state)) {
-    case _RAISE:
-    case _LOWER:
-    case _ADJUST:
-        rgbflag(0x00,  0x00, 0xFF);
-        break;
-    case _MOVE:
-    case _MOUSE:
-        rgbflag(0xFF,  0x00, 0x00);
-        break;
-    case _CMD:
-        rgbflag(0x00,  0xFF, 0x00);
-        break;
-    case _EDIT:
-        rgbflag(0x7A,  0x00, 0xFF);
-        break;
-    default: //  for any other layers, or the default layer
-        rgbflag(0x00,  0xFF, 0xFF);
-        break;
-    }
-  return state;
 }
